+++ example/lib/src/candlestick_chart.dart
@@ -108,6 +108,8 @@ class CandlestickChartExample extends StatelessWidget {
         SetupDivider(),
         _AxisStyleSetupGroup(),
         SetupDivider(),
+        _ZeroLineStyleSetupGroup(),
+        SetupDivider(),
         _DropLineStyleSetupGroup(),
         SetupDivider(),
         _CandleStickStyleSetupGroup(),
@@ -369,6 +371,13 @@ class _GeneralSettingsSetupGroup extends StatelessWidget {
                 _settings.add(data.copyWith(yAxisLabelSpacing: doubleValue));
               },
               title: const Text('yAxisLabelSpacing'),
+            ),
+            CheckboxListTile(
+              controlAffinity: ListTileControlAffinity.trailing,
+              value: data.showZeroLine,
+              onChanged: (value) =>
+                  _settings.add(data.copyWith(showZeroLine: value == true)),
+              title: const Text('showZeroLine'),
             ),
             CheckboxListTile(
               controlAffinity: ListTileControlAffinity.trailing,
@@ -642,6 +651,53 @@ class _AxisStyleSetupGroup extends StatelessWidget {
                 );
               },
               title: const Text('xAxisSelectedLabelBorderRadius'),
+            ),
+          ],
+        );
+      },
+    );
+  }
+}
+
+class _ZeroLineStyleSetupGroup extends StatelessWidget {
+  const _ZeroLineStyleSetupGroup();
+
+  @override
+  Widget build(BuildContext context) {
+    return StreamBuilder<CandlestickChartStyle>(
+      stream: _style,
+      initialData: _style.value,
+      builder: (context, style) {
+        final data = style.requireData;
+
+        return SetupGroup(
+          title: 'Zero Line Style',
+          children: [
+            IntListTile(
+              value: data.zeroLineStyle.stroke.toInt(),
+              onChanged: (value) {
+                _style.add(
+                  data.copyWith(
+                    zeroLineStyle: data.zeroLineStyle.copyWith(
+                      stroke: value.toDouble(),
+                    ),
+                  ),
+                );
+              },
+              title: const Text('stroke'),
+            ),
+            ColorListTile(
+              value: data.zeroLineStyle.color,
+              onChanged: (value) {
+                _style.add(
+                  data.copyWith(
+                    zeroLineStyle: data.zeroLineStyle.copyWith(
+                      color: value,
+                    ),
+                  ),
+                );
+              },
+              title: const Text('color'),
             ),
           ],
         );